+++ src/main/java/com/sdl/selenium/utils/browsers/FirefoxConfigReader.java
@@ -110,7 +110,6 @@ public class FirefoxConfigReader extends AbstractBrowserConfigReader {
                 setProfilePreferences(profile);
             }
         }
-        setUploadPathExe();
         return profile;
     }